+++ src/main/java/com/tsoft/civilization/civilization/tile/TileSupplyStrategy.java
@@ -6,13 +6,13 @@ import java.util.Comparator;
 
 public enum TileSupplyStrategy {
 
-    MAX_FOOD(Comparator.comparingInt(Supply::getFood)),
-    MAX_PRODUCTION(Comparator.comparingInt(Supply::getProduction)),
-    MAX_GOLD(Comparator.comparingInt(Supply::getGold)),
-    MAX_SCIENCE(Comparator.comparingInt(Supply::getScience)),
-    MAX_CULTURE(Comparator.comparingInt(Supply::getCulture)),
-    MAX_FAITH(Comparator.comparingInt(Supply::getFaith)),
-    MAX_GREAT_PERSON(Comparator.comparingInt(Supply::getGreatPerson));
+    MAX_FOOD(Comparator.comparingDouble(Supply::getFood)),
+    MAX_PRODUCTION(Comparator.comparingDouble(Supply::getProduction)),
+    MAX_GOLD(Comparator.comparingDouble(Supply::getGold)),
+    MAX_SCIENCE(Comparator.comparingDouble(Supply::getScience)),
+    MAX_CULTURE(Comparator.comparingDouble(Supply::getCulture)),
+    MAX_FAITH(Comparator.comparingDouble(Supply::getFaith)),
+    MAX_GREAT_PERSON(Comparator.comparingDouble(Supply::getGreatPerson));
 
     private final Comparator<Supply> comparator;
